spot route finished

+++ backend/db/models/spot.js
@@ -21,7 +21,6 @@ module.exports = (sequelize, DataTypes) => {
         foreignKey: "spotId",
         onDelete: "CASCADE",
         hooks: true,
-        // as: "previewImage",
       });
       Spot.hasMany(models.Booking, {
         foreignKey: "spotId",
@@ -55,7 +54,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
       },
       lat: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: false,
         validate: {
           min: -90,
@@ -63,7 +62,7 @@ module.exports = (sequelize, DataTypes) => {
         },
       },
       lng: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: false,
         validate: {
           min: -180,
@@ -79,7 +78,7 @@ module.exports = (sequelize, DataTypes) => {
         allowNull: false,
       },
       price: {
-        type: DataTypes.DECIMAL,
+        type: DataTypes.FLOAT,
         allowNull: false,
       },
     },
